Biter Pet - Initial status label work
Added some extra status's, but more in the TODO comment.
Not documented any of the new statusMessage tables as these are still in flux. The other new options I've added are documented.
Made all of the variables in the function start with an underscore to help spot mistakes in scoping.